@@ -15,33 +15,40 @@
     "typecheck": "tsc"
   },
   "dependencies": {
-    "@docusaurus/core": "0.0.0-5396",
-    "@docusaurus/plugin-ideal-image": "0.0.0-5396",
-    "@docusaurus/preset-classic": "0.0.0-5396",
-    "@docusaurus/remark-plugin-npm2yarn": "^2.2.0",
-    "@fortawesome/fontawesome-svg-core": "^6.2.1",
-    "@fortawesome/free-brands-svg-icons": "^6.2.1",
-    "@fortawesome/free-regular-svg-icons": "^6.2.1",
-    "@fortawesome/free-solid-svg-icons": "^6.2.1",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/plugin-ideal-image": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
+    "@docusaurus/remark-plugin-npm2yarn": "^3.0.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.2",
+    "@fortawesome/free-brands-svg-icons": "^6.4.2",
+    "@fortawesome/free-regular-svg-icons": "^6.4.2",
+    "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@mdx-js/react": "^1.6.22",
-    "@popperjs/core": "^2.11.6",
-    "@types/react-dom": "^18.0.9",
-    "clsx": "^1.2.1",
+    "@mdx-js/react": "^3.0.0",
+    "@popperjs/core": "^2.11.8",
+    "@swc/core": "^1.3.96",
+    "@types/react-dom": "^18.2.15",
+    "clsx": "^2.0.0",
     "docusaurus-plugin-plausible": "^0.0.5",
-    "docusaurus-plugin-sass": "^0.2.2",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "docusaurus-plugin-sass": "^0.2.5",
+    "docusaurus-remark-plugin-tab-blocks": "^2.0.0",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-ideal-image": "^0.0.5",
     "react-popper": "^2.3.0",
-    "sass": "^1.56.1"
+    "sass": "^1.69.5",
+    "swc-loader": "^0.2.3"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "0.0.0-5396",
-    "@tsconfig/docusaurus": "^1.0.5",
-    "jest": "^29.3.1",
-    "typescript": "^4.7.4"
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@docusaurus/tsconfig": "^3.0.0",
+    "@types/react": "^18.2.37",
+    "jest": "^29.7.0",
+    "prop-types": "^15.8.1",
+    "react-waypoint": "^10.3.0",
+    "typescript": "^5.2.2",
+    "webpack": "^5.89.0"
   },
   "browserslist": {
     "production": [
@@ -57,5 +64,6 @@
   },
   "engines": {
     "node": ">=16.14"
-  }
+  },
+  "packageManager": "yarn@4.0.1"
 }
